cancer-ai-platform/backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# AI / ML
torch
torchvision
tensorflow
numpy
opencv-python-headless
Pillow
scipy
scikit-learn

# RAG + Gemini
google-generativeai
chromadb
sentence-transformers

# Utilities
loguru>=0.7.0
aiofiles>=23.0.0
